requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.8
django-browser-reload==1.8.0
django-colorfield==0.8.0
django-htmx==1.14.0
idna==3.4
inflection==0.5.1
Levenshtein==0.21.0
nameparser==1.1.2
Pillow==9.5.0
python-Levenshtein==0.21.0
PyYAML==6.0
rapidfuzz==3.0.0
requests==2.29.0
slippers==0.5.0
soupsieve==2.4.1
sqlparse==0.4.4
thefuzz==0.19.0
urllib3==1.26.15
whitenoise==6.4.0
